refactor: update infra
