+++ src/infra/criptography/jwt-adapter.ts
@@ -1,7 +1,7 @@
 import jwt from 'jsonwebtoken';
-import { IEncrypter, IDecrypter, IPayload } from '@/use-cases/interfaces';
+import { IEncrypter, IDecrypter, IPayload } from '@/use-cases/user/interfaces';
 import { Either, success, error } from '@/shared';
-import { UnauthorizedError } from '@/use-cases/errors';
+import { UnauthorizedError } from '@/use-cases/user/errors';
 
 export class JwtAdapter implements IEncrypter, IDecrypter {
   constructor(
