+++ src/infra/universally-unique-identifier/index.ts
@@ -1,0 +1,1 @@
+export * from '@/infra/universally-unique-identifier/universally-unique-identifier-adapter';
