+++ src/infra/universally-unique-identifier/universally-unique-identifier-adapter.ts
@@ -1,5 +1,5 @@
 import { v4 as uuid } from 'uuid';
-import { IIdGenerator } from '@/use-cases/interfaces';
+import { IIdGenerator } from '@/use-cases/user/interfaces';
 
 export class UuidAdapter implements IIdGenerator {
   async generate(): Promise<string> {
